Test cache for a cargo bin project
This is an end executable project so that we have Cargo.lock
checked in. Its checksum is used to hit the registry cache.
If the registry cache is a miss, we still want the registry index
restored from a smaller cache, with fallback to its earlier snapshots,
so that cargo spends less time synchronizing the index.